refactor(v5): altera estrutura de pastas
