+++ src/app/components/ui/table.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import { cn } from '@/lib/utils'
+import { cn } from '@/app/lib/utils'
 
 const Table = React.forwardRef<
 	HTMLTableElement,
