+++ src/app/components/ui/toast.tsx
@@ -5,7 +5,7 @@ import { type VariantProps, cva } from 'class-variance-authority'
 import { X } from 'lucide-react'
 import * as React from 'react'
 
-import { cn } from '@/lib/utils'
+import { cn } from '@/app/lib/utils'
 
 const ToastProvider = ToastPrimitives.Provider
