+++ src/app/components/ui/toaster.tsx
@@ -7,8 +7,8 @@ import {
 	ToastProvider,
 	ToastTitle,
 	ToastViewport,
-} from '@/components/ui/toast'
-import { useToast } from '@/hooks/use-toast'
+} from '@/app/components/ui/toast'
+import { useToast } from '@/app/hooks/use-toast'
 
 export function Toaster() {
 	const { toasts } = useToast()
