+++ src/app/error.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { Button } from '@/components/ui/button'
+import { Button } from '@/app/components/ui/button'
 import { useEffect } from 'react'
 
 const Error = ({
