+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from 'next'
 import localFont from 'next/font/local'
 import './globals.css'
-import { Toaster } from '@/components/ui/toaster'
-import { ReactQueryProvider } from '@/lib/react-query-provider'
+import { Toaster } from '@/app/components/ui/toaster'
+import { ReactQueryProvider } from '@/app/lib/react-query-provider'
 import { Analytics } from '@vercel/analytics/react'
 import { SpeedInsights } from '@vercel/speed-insights/next'
